demo/requirements.txt
gradio
scipy
numpy
Pillow
opencv-python-headless
